Port over HTML from previous app
Originaly when I generated the app, I did so with no flags and created
your normal HTML app. However, I realized that is not what I wanted to
do. I wanted to create the app using the --no-html flag so it would be a
json app (done in a previous commit).

There would be a couple of HTML pages in the app so I ported over the
boilerplate code from the other generated app, adding in files that did
not exist and merging into files that already did.

All tests are green including the added in html tests.