Remove all actions that aren't index or show
I think this will be temporary. For now, only going to have index and
show for the purposes of people accessing the information in the API. I
may later add back in the ability to create, edit, update, or delete
resources. Since this would be released with all of the necessary
information and changes will not be that often, going to keep it simple
for now.

Future plans would have things able to be created, edited, updated, or
deleted with an API key which had the proper permissions. Again, not
really needed for the purposes of things as they currently stand.